# backend/main-fiber/docs/swagger.yaml
basePath: /api
definitions:
...
  api.OrgUnitCreate:
    properties:
      name:
        additionalProperties:
          type: string
        description: '{en:"Computer Engineering"}'
        type: object
      parent_path:
        type: string
      path:
        description: '"/faculty/eng/com"'
        type: string
      short_name:
        additionalProperties:
          type: string
        description: '{en:"COM", th:"คอม"}'
        type: object
      slug:
        type: string
      sort:
        type: integer
      status:
        description: default "active"
        type: string
      type:
        description: '"faculty"|"program"|"club"|...'
        type: string
      visibility:
        description: default "public"
        type: string
    type: object
  api.OrgUnitDoc:
    properties:
      ancestors:
        items:
          type: string
        type: array
      created_at:
        type: string
      depth:
        type: integer
      name:
        additionalProperties:
          type: string
        type: object
      parent_path:
        type: string
      path:
        type: string
      short_name:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      sort:
        type: integer
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
      visibility:
        type: string
    type: object
  api.OrgUnitUpdate:
    properties:
      name:
        additionalProperties:
          type: string
        type: object
      short_name:
        additionalProperties:
          type: string
        type: object
      slug:
        type: string
      sort:
        type: integer
      status:
        type: string
      visibility:
        type: string
    type: object
...
  api.Policy:
    properties:
      actions:
        items:
          type: string
        type: array
      created_at:
        type: string
      effect:
        description: '"allow"'
        type: string
      enabled:
        type: boolean
      key:
        type: string
      position_key:
        description: e.g. "head"
        type: string
      scope:
        description: '"exact" | "subtree"'
        type: string
      where:
        $ref: '#/definitions/api.PolicyWhere'
    type: object
  api.PolicyWhere:
    properties:
      org_prefix:
        description: e.g. "/faculty/"
        type: string
    type: object
...
  api.abilitiesResp:
    properties:
      abilities:
        additionalProperties:
          type: boolean
        type: object
      org_path:
        type: string
      version:
        type: string
    type: object
...
info:
  contact: {}
  description: This is my API
  title: My API
  version: "1.0"
paths:
  /api/abilities:
    get:
      consumes:
      - application/json
      description: Returns allowed actions for the user in the specified org path
      parameters:
      - description: Organization path
        in: query
        name: org_path
        required: true
        type: string
      - description: Comma-separated actions to check
        in: query
        name: actions
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.abilitiesResp'
      summary: Get user abilities for an org path
      tags:
      - abilities
  /api/abilities/where:
    get:
      consumes:
      - application/json
      description: Returns org paths where the user has permission for the specified
        action
      parameters:
      - description: Action to check
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List org paths where user can perform an action
      tags:
      - abilities
...
  /api/org/units/node:
    delete:
      description: Delete organization unit node by path (leaf-only)
      parameters:
      - description: Org unit path
        in: query
        name: path
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete org unit
      tags:
      - org_units
    get:
      consumes:
      - application/json
      description: Get organization unit node by path
      parameters:
      - description: Org unit path
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OrgUnitDoc'
      summary: Get org unit
      tags:
      - org_units
    patch:
      consumes:
      - application/json
      description: Update organization unit node by path
      parameters:
      - description: Org unit path
        in: query
        name: path
        required: true
        type: string
      - description: Fields to update
        in: body
        name: OrgUnitUpdate
        required: true
        schema:
          $ref: '#/definitions/api.OrgUnitUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OrgUnitDoc'
      summary: Update org unit
      tags:
      - org_units
    post:
      consumes:
      - application/json
      description: Create a new organization unit node
      parameters:
      - description: Org unit info
        in: body
        name: OrgUnitCreate
        required: true
        schema:
          $ref: '#/definitions/api.OrgUnitCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.OrgUnitDoc'
      summary: Create org unit
      tags:
      - org_units
...
  /policies:
    delete:
      description: Delete policies by org_prefix and optional position_key
      parameters:
      - description: Organization prefix
        in: query
        name: org_prefix
        required: true
        type: string
      - description: Position key
        in: query
        name: position_key
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete policy
      tags:
      - policies
    get:
      consumes:
      - application/json
      description: List policies with optional filters
      parameters:
      - description: Organization prefix
        in: query
        name: org_prefix
        type: string
      - description: Position key
        in: query
        name: position_key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.Policy'
            type: array
        "500":
          description: DB error or decode error
          schema:
            type: string
      summary: List policies
      tags:
      - policies
    post:
      consumes:
      - application/json
      description: Create a new policy
      parameters:
      - description: Policy info
        in: body
        name: Policy
        required: true
        schema:
          $ref: '#/definitions/api.Policy'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.Policy'
        "400":
          description: Invalid body or missing fields
          schema:
            type: string
        "500":
          description: Insert failed
          schema:
            type: string
      summary: Create policy
      tags:
      - policies
    put:
      consumes:
      - application/json
      description: Upsert (create or update) a policy by position_key, org_prefix,
        and scope
      parameters:
      - description: Policy info
        in: body
        name: Policy
        required: true
        schema:
          $ref: '#/definitions/api.Policy'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Policy'
      summary: Upsert policy
      tags:
      - policies
...
swagger: "2.0"